# docs/swagger/swagger.yaml
definitions:
...
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /sessions/{sessionId}/media/clear-cache:
    post:
      consumes:
      - application/json
      description: Clear all cached media files for the session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Clear media cache
      tags:
      - Media
  /sessions/{sessionId}/media/download:
    post:
      consumes:
      - application/json
      description: Download media file from WhatsApp message
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Download media from WhatsApp
      tags:
      - Media
  /sessions/{sessionId}/media/info:
    get:
      description: Get information about media files
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Get media information
      tags:
      - Media
  /sessions/{sessionId}/media/list:
    get:
      description: List all cached media files for the session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: List cached media files
      tags:
      - Media
  /sessions/{sessionId}/media/stats:
    get:
      description: Get statistics about media usage for the session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Get media statistics
      tags:
      - Media
...
swagger: "2.0"